support ranges of host ports to be mapped to a container port
