+++ tests/unit/test_dockerclient.py
@@ -7,6 +7,7 @@ from unittest.mock import patch
 import pytest
 
 from localstack import config
+from localstack.utils.bootstrap import extract_port_flags
 from localstack.utils.docker_utils import CmdDockerClient, DockerContainerStatus, PortMappings, Util
 
 LOG = logging.getLogger(__name__)
@@ -54,17 +55,16 @@ class TestDockerClient(unittest.TestCase):
 def test_argument_parsing():
     test_port_string = "-p 80:8080/udp"
     test_port_string_with_host = "-p 127.0.0.1:6000:7000/tcp"
+    test_port_string_many_to_one = "-p 9230-9231:9230"
     test_env_string = "-e TEST_ENV_VAR=test_string=123"
     test_mount_string = "-v /var/test:/opt/test"
-    argument_string = (
-        f"{test_port_string} {test_env_string} {test_mount_string} {test_port_string_with_host}"
-    )
+    argument_string = f"{test_port_string} {test_env_string} {test_mount_string} {test_port_string_with_host} {test_port_string_many_to_one}"
     env_vars = {}
     ports = PortMappings()
     mounts = []
     Util.parse_additional_flags(argument_string, env_vars, ports, mounts)
     assert env_vars == {"TEST_ENV_VAR": "test_string=123"}
-    assert ports.to_str() == "-p 80:8080/udp -p 6000:7000"
+    assert ports.to_str() == "-p 80:8080/udp -p 6000:7000 -p 9230-9231:9230"
     assert mounts == [("/var/test", "/opt/test")]
     argument_string = (
         "--add-host host.docker.internal:host-gateway --add-host arbitrary.host:127.0.0.1"
@@ -105,6 +105,61 @@ def test_argument_parsing():
     argument_string = r'-v "/tmp/test.jar:/tmp/foo bar/test.jar" --network mynet123'
     _, _, _, _, network = Util.parse_additional_flags(argument_string)
     assert network == "mynet123"
+
+
+class TestCommandLine(unittest.TestCase):
+    def test_extract_port_flags(self):
+        port_mappings = PortMappings()
+        flags = extract_port_flags("foo -p 1234:1234 bar", port_mappings=port_mappings)
+        self.assertEqual("foo  bar", flags)
+        mapping_str = port_mappings.to_str()
+        self.assertEqual("-p 1234:1234", mapping_str)
+
+        port_mappings = PortMappings()
+        flags = extract_port_flags(
+            "foo -p 1234:1234 bar -p 80-90:81-91 baz", port_mappings=port_mappings
+        )
+        self.assertEqual("foo  bar  baz", flags)
+        mapping_str = port_mappings.to_str()
+        self.assertIn("-p 1234:1234", mapping_str)
+        self.assertIn("-p 80-90:81-91", mapping_str)
+
+    def test_overlapping_port_ranges(self):
+        port_mappings = PortMappings()
+        port_mappings.add(4590)
+        port_mappings.add(4591)
+        port_mappings.add(4593)
+        port_mappings.add(4592)
+        port_mappings.add(4593)
+        result = port_mappings.to_str()
+        # assert that ranges are non-overlapping, i.e., no duplicate ports
+        self.assertEqual("-p 4590-4592:4590-4592 -p 4593:4593", result)
+
+    def test_port_ranges_with_bind_host(self):
+        port_mappings = PortMappings(bind_host="0.0.0.0")
+        port_mappings.add(5000)
+        port_mappings.add(5001)
+        port_mappings.add(5003)
+        port_mappings.add([5004, 5006], 9000)
+        result = port_mappings.to_str()
+        self.assertEqual(
+            "-p 0.0.0.0:5000-5001:5000-5001 -p 0.0.0.0:5003:5003 -p 0.0.0.0:5004-5006:9000", result
+        )
+
+    def test_port_ranges_with_bind_host_to_dict(self):
+        port_mappings = PortMappings(bind_host="0.0.0.0")
+        port_mappings.add(5000, 6000)
+        port_mappings.add(5001, 7000)
+        port_mappings.add(5003, 8000)
+        port_mappings.add([5004, 5006], 9000)
+        result = port_mappings.to_dict()
+        expected_result = {
+            "6000/tcp": ("0.0.0.0", 5000),
+            "7000/tcp": ("0.0.0.0", 5001),
+            "8000/tcp": ("0.0.0.0", 5003),
+            "9000/tcp": ("0.0.0.0", [5004, 5005, 5006]),
+        }
+        self.assertEqual(expected_result, result)
 
 
 def list_in(a, b):
